some valgrind testing and updated the test script

+++ tests/tests.sh
@@ -1,18 +1,23 @@
 #!/bin/bash
 
 roms="$HOME/Projects/LLD_gbemu/roms/*"
-
+prog_name="main"
 MAX_WAIT_TIME=30
+
+function rm_s()
+{
+	if [ -f $1 ]
+	then
+		rm $1
+	fi
+}
 
 for rom in $roms
 do
 	if [[ $rom =~ roms/[0-9] ]]
 	then
-		if [ -f nohup.out ]
-		then
-			rm nohup.out
-		fi
-		nohup ./main "$rom" 2>/dev/null &
+		rm_s nohup.out
+		nohup ./$prog_name "$rom" 2>/dev/null &
 		sleep 1
 		time=0
 		while ! tail nohup.out | grep -q Passed
@@ -24,12 +29,12 @@ do
 				echo "$rom Failed"
 			fi
 		done
-		kill -9 $!
-		wait $! 2>/dev/null
+		if ps -a | grep -q $prog_name
+		then
+			kill -9 $!
+			wait $! 2>/dev/null
+		fi
 	fi
 done
 
-if [ -f nohup.out ]
-then
-	rm nohup.out
-fi
+rm_s nohup.out
